requirements.txt
langchain
langchain_google_genai
langchain-community
langchain-core
python-dotenv
streamlit
pinecone
pypdf
langchain_pinecone
PyPDF2
FAISS